Major Updates
These are mostly a rewrite of the banker code to be more efficient and
to reuse code over and over again instead of having to change every
place where it is used.  This maintains a feeling of consistency when it
is run. When making deposits and withdrawals the Account Number must be
entered each time.  There is a option to get a list of account numbers
but it still must be entered.
I need to work on this to make it better. I also need to work on the EOM
as I think it is too cumbersome and need to make it more efficient. I
also added an Excel file to help me with the calculations to make sure
it is working properly. I expanded the test data a little to make it a
better fit completely testing!