Update Banker
This will be the working copy of Banker and I have included some more
test data.  I have checked thoroughly the calculations and the EOM
processing and they are working perfectly. Awny sent the update to the
code challenge we were having for FX and it works nice. I have included
it here.