Up to Date
Ok I believe the code is ready. I thoroughly tested it and made some
minor corrections and I added a option to ask the user for the bank name
when it is first run. This will make it customizable completely for the
Banker (Text based side)!  I removed FXBanker from the application as we
will not be using this version. I will add the new version when I get it
more complete. For now it will be called NewFXBanker.java
If you don't have this one then it will need to be added to your eclipse
code