Update new Banker
Updated the new Banker and got the account to work properly.  Working on
fixing a few minor bugs. I can implement the account adding for gold and
regular now but I want to make sure the code is complete before I do. I
updated the customer form and made it color coordinated with the account
form