Update of Phase1
This is the update to phase 1 that will only let you remove a customer
if they have no accounts. I also changed the Transaction constructor
that the remove customer method calls to include the customerID and the
amount it shows to 0.00. In the Customer class I added a private field
called active and it is true or false. When an account is added, this is
updated to true and when trying to remove a customer it only shows those
customers that have no accounts. It processes each customer looking
through the accounts and if it finds a match it sets the active flag to
true and false if it finds none.