Phase 2
This is the official start of phase 2. Phase 1 is available in the
folder BankPhase1 but all code from here for phase 2 will be in the
folder BankPhase2. You will need to update eclipse source files
accordingly. If there are any questions please text me before making any
changes the code!