Update
I updated Phase 1 to include file save and load when starting and
exiting the program.  I also added auto Customer and Account number
generation. This needs to be thoroughly test now to see if it will ever
break.
I also worked on the Phase 2 and it still needs some work but it works
ok.